requirements.txt
# LMS FinAI Service (Merged) - Dependencies
# Includes: new roinet-lms-finai + legacy lms-finai-service_v17 APIs

# FastAPI framework
fastapi==0.128.0
uvicorn[standard]==0.40.0
starlette==0.50.0
anyio==4.12.0

# Pydantic
pydantic==2.12.5
pydantic-settings==2.12.0
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP client
httpx==0.28.1
httpcore==1.0.9

# JWT validation
python-jose[cryptography]==3.3.0
cryptography>=41.0.0

# OpenAI
openai==2.14.0
jiter==0.12.0

# Knowledge base / documents
pypdf==6.5.0

# Redis
redis==7.1.0

# Utils
python-dotenv==1.2.1
python-multipart==0.0.21
orjson==3.11.5
tenacity==9.1.2
tqdm==4.67.1
apscheduler==3.10.4

# CSV processing
pandas==2.1.4
numpy

# Excel
openpyxl

# Logging
structlog==25.5.0

# WebSocket + templating (telephony streaming + dynamic prompts)
websockets==12.0
Jinja2==3.1.4

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
